Remove helper functions that interact with Context from Tracer and move into contrib
