+++ api/src/main/java/io/opentelemetry/trace/Tracer.java
@@ -21,7 +21,6 @@ import io.opentelemetry.context.Scope;
 import io.opentelemetry.context.propagation.BinaryFormat;
 import io.opentelemetry.context.propagation.HttpTextFormat;
 import io.opentelemetry.resource.Resource;
-import java.util.concurrent.Callable;
 import javax.annotation.Nullable;
 
 /**
@@ -140,133 +139,6 @@ public interface Tracer {
    */
   @MustBeClosed
   Scope withSpan(Span span);
-
-  /**
-   * Returns a {@link Runnable} that runs the given task with the given {@code Span} in the current
-   * context.
-   *
-   * <p>Users may consider to use {@link Span.Builder#startSpanAndRun(Runnable)}.
-   *
-   * <p>Any error will end up as a {@link Status#UNKNOWN}.
-   *
-   * <p>IMPORTANT: Caller must manually propagate the entire {@code io.grpc.Context} when wraps a
-   * {@code Runnable}, see the examples.
-   *
-   * <p>IMPORTANT: Caller must manually end the {@code Span} within the {@code Runnable}, or after
-   * the {@code Runnable} is executed.
-   *
-   * <p>Example with Executor wrapped with {@link io.grpc.Context#currentContextExecutor}:
-   *
-   * <pre><code>
-   * class MyClass {
-   *   private static Tracer tracer = Trace.getTracer();
-   *   void handleRequest(Executor executor) {
-   *     Span span = tracer.spanBuilder("MyRunnableSpan").startSpan();
-   *     executor.execute(tracer.withSpan(span, new Runnable() {
-   *      {@literal @}Override
-   *       public void run() {
-   *         try {
-   *           sendResult();
-   *         } finally {
-   *           span.end();
-   *         }
-   *       }
-   *     }));
-   *   }
-   * }
-   * </code></pre>
-   *
-   * <p>Example without Executor wrapped with {@link io.grpc.Context#currentContextExecutor}:
-   *
-   * <pre><code>
-   * class MyClass {
-   *   private static Tracer tracer = Trace.getTracer();
-   *   void handleRequest(Executor executor) {
-   *     Span span = tracer.spanBuilder("MyRunnableSpan").startSpan();
-   *     executor.execute(Context.wrap(tracer.withSpan(span, new Runnable() {
-   *      {@literal @}Override
-   *       public void run() {
-   *         try {
-   *           sendResult();
-   *         } finally {
-   *           span.end();
-   *         }
-   *       }
-   *     })));
-   *   }
-   * }
-   * </code></pre>
-   *
-   * @param span the {@code Span} to be set as current.
-   * @param runnable the {@code Runnable} to withSpan in the {@code Span}.
-   * @return the {@code Runnable}.
-   * @since 0.1.0
-   */
-  Runnable withSpan(Span span, Runnable runnable);
-
-  /**
-   * Returns a {@link Callable} that runs the given task with the given {@code Span} in the current
-   * context.
-   *
-   * <p>Users may consider to use {@link Span.Builder#startSpanAndCall(Callable)}.
-   *
-   * <p>Any error will end up as a {@link Status#UNKNOWN}.
-   *
-   * <p>IMPORTANT: Caller must manually propagate the entire {@code io.grpc.Context} when wraps a
-   * {@code Callable}, see the examples.
-   *
-   * <p>IMPORTANT: Caller must manually end the {@code Span} within the {@code Callable}, or after
-   * the {@code Callable} is executed.
-   *
-   * <p>Example with Executor wrapped with {@link io.grpc.Context#currentContextExecutor}:
-   *
-   * <pre><code>
-   * class MyClass {
-   *   private static Tracer tracer = Trace.getTracer();
-   *   void handleRequest(Executor executor) {
-   *     Span span = tracer.spanBuilder("MyRunnableSpan").startSpan();
-   *     executor.execute(tracer.withSpan(span, new Callable&lt;MyResult&gt;() {
-   *      {@literal @}Override
-   *       public MyResult call() throws Exception {
-   *         try {
-   *           return sendResult();
-   *         } finally {
-   *           span.end();
-   *         }
-   *       }
-   *     }));
-   *   }
-   * }
-   * </code></pre>
-   *
-   * <p>Example without Executor wrapped with {@link io.grpc.Context#currentContextExecutor}:
-   *
-   * <pre><code>
-   * class MyClass {
-   *   private static Tracer tracer = Trace.getTracer();
-   *   void handleRequest(Executor executor) {
-   *     Span span = tracer.spanBuilder("MyRunnableSpan").startSpan();
-   *     executor.execute(Context.wrap(tracer.withSpan(span, new Callable&lt;MyResult&gt;() {
-   *      {@literal @}Override
-   *       public MyResult call() throws Exception {
-   *         try {
-   *           return sendResult();
-   *         } finally {
-   *           span.end();
-   *         }
-   *       }
-   *     })));
-   *   }
-   * }
-   * </code></pre>
-   *
-   * @param span the {@code Span} to be set as current.
-   * @param callable the {@code Callable} to run in the {@code Span}.
-   * @param <V> the result type of method call.
-   * @return the {@code Callable}.
-   * @since 0.1.0
-   */
-  <V> Callable<V> withSpan(Span span, final Callable<V> callable);
 
   /**
    * Returns a {@link Span.Builder} to create and start a new child {@link Span} as a child of to
