+++ contrib/src/main/java/io/opentelemetry/contrib/trace/CurrentSpanUtils.java
@@ -14,10 +14,9 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.sdk.trace.internal;
+package io.opentelemetry.contrib.trace;
 
 import io.grpc.Context;
-import io.opentelemetry.context.Scope;
 import io.opentelemetry.trace.Span;
 import io.opentelemetry.trace.Status;
 import io.opentelemetry.trace.unsafe.ContextUtils;
@@ -27,29 +26,6 @@ import java.util.concurrent.Callable;
 public final class CurrentSpanUtils {
   // No instance of this class.
   private CurrentSpanUtils() {}
-
-  /**
-   * Returns The {@link Span} from the current context.
-   *
-   * @return The {@code Span} from the current context.
-   */
-  public static Span getCurrentSpan() {
-    return ContextUtils.getValue(Context.current());
-  }
-
-  /**
-   * Enters the scope of code where the given {@link Span} is in the current context, and returns an
-   * object that represents that scope. The scope is exited when the returned object is closed.
-   *
-   * <p>Supports try-with-resource idiom.
-   *
-   * @param span The {@code Span} to be set to the current context.
-   * @return An object that defines a scope where the given {@code Span} is set to the current
-   *     context.
-   */
-  public static Scope withSpan(Span span) {
-    return new ScopeInSpan(span);
-  }
 
   /**
    * Wraps a {@link Runnable} so that it executes with the {@code span} as the current {@code Span}.
@@ -75,27 +51,7 @@ public final class CurrentSpanUtils {
     return new CallableInSpan<C>(span, callable, endSpan);
   }
 
-  // Defines an arbitrary scope of code as a traceable operation. Supports try-with-resources idiom.
-  private static final class ScopeInSpan implements Scope {
-    private final Context origContext;
-
-    /**
-     * Constructs a new {@link ScopeInSpan}.
-     *
-     * @param span is the {@code Span} to be added to the current {@code io.grpc.Context}.
-     */
-    private ScopeInSpan(Span span) {
-      origContext = ContextUtils.withValue(Context.current(), span).attach();
-    }
-
-    @Override
-    public void close() {
-      Context.current().detach(origContext);
-    }
-  }
-
   private static final class RunnableInSpan implements Runnable {
-    // TODO(bdrutu): Investigate if the extra private visibility increases the generated bytecode.
     private final Span span;
     private final Runnable runnable;
     private final boolean endSpan;
